# extend CONTINUATION_ALIGN_STYLE=FIXED for space indentation