--- a/yapftests/yapf_test.py
+++ b/yapftests/yapf_test.py
@@ -1309,6 +1309,33 @@ CONTINUATION_ALIGN_STYLE = valign-right
           expected_formatted_code,
           extra_options=['--style={0}'.format(stylepath)])
 
+  def testUseSpacesContinuationAlignStyleFixed(self):
+    unformatted_code = """\
+def foo_function(arg1, arg2, arg3):
+  return ['hello', 'world',]
+"""
+    expected_formatted_code = """\
+def foo_function(arg1, arg2,
+        arg3):
+    return [
+            'hello',
+            'world',
+    ]
+"""
+    style_contents = u"""\
+[style]
+based_on_style = chromium
+COLUMN_LIMIT=32
+INDENT_WIDTH=4
+CONTINUATION_INDENT_WIDTH=8
+CONTINUATION_ALIGN_STYLE = fixed
+"""
+    with utils.TempFileContents(self.test_tmpdir, style_contents) as stylepath:
+      self.assertYapfReformats(
+          unformatted_code,
+          expected_formatted_code,
+          extra_options=['--style={0}'.format(stylepath)])
+
   def testStyleOutputRoundTrip(self):
     unformatted_code = textwrap.dedent("""\
         def foo_function():