This PR make `CONTINUATION_ALIGN_STYLE=FIXED` option available for indenting with space.

Ref: #769