@andreasvc 2019-10-24
Great!

It adds an unexpected linebreak before the return type annotation though:
```python
def _as_prolog_str_full(                                                                          
        value: Union[PlainOldData, List[Any], bool, str, int, Dict[Any, Any],                             None, Exception]                                                                          
) -> str:                                                                                         
    """Recursively turn an object into a Prolog term.""" 
```
I expected the ``) -> str:``` to go at the end of the previous line; if a line break needs to be inserted (which here it doesn't), I expect it to have the same amount of continuation indentation, because the line is still being continued.