@kamahen 2019-10-24
This doesn't give the extra 4 spaces of indent on a continuation that I was hoping for. :(

(I also saw the extra indent before ") -> str", which wasn't there before)