@andreasvc 2019-10-24
@kamahen For me the extra continuation indent works; what's the example you run it on, and what are your settings?
@kamahen I'm confused, you saw an extra indent before ``) -> str``? I'm seeing the *lack* of any indentation on that line as shown in my example...